pyproject.toml
[project]
name = "chatgpt-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "graphviz>=0.21",
    "openai-agents[viz]==0.4.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
